+++ packages/utls/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @clancyapp/utls
 
+## 0.0.6
+
 ## 0.0.5
 
 ### Patch Changes