+++ packages/utls/platforms/darwin-x64/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @clancyapp/utls-darwin-x64
 
+## 0.0.6
+
 ## 0.0.5
 
 ### Patch Changes