+++ packages/utls/platforms/linux-arm64/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @clancyapp/utls-linux-arm64
 
+## 0.0.6
+
 ## 0.0.5
 
 ### Patch Changes